logout/in, autorisation, user listing relationships
Made logout/login links, functions and routes , set authorisation for users and connected them to their listings

+++ routes/web.php
@@ -24,33 +24,44 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 //show create form 
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware("auth");//only user create listing ;
 
 
 // Store Listing Data
-Route::POST('/listings', [ListingController::class, 'store']);
+Route::POST('/listings', [ListingController::class, 'store'])->middleware("auth");//only user create listing ;
 
 //Show edit form
 Route::get('/listings/{listing}/edit', 
-[ListingController::class, 'edit']);
+[ListingController::class, 'edit'])->middleware("auth");//only user edit listing ;
 
 //Edit, submit to update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware("auth");//only user update listing ;
 
 //Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware("auth");//only user delete listing ;
 
 
 
 // Single Listing
-Route::get('/listings/{listing}', [ListingController::class, 'show']);
+Route::get('/listings/{listing}', [ListingController::class, 'show'])->middleware("guest");
 
 //Show register create form 
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware("guest");//guest can access registration
 
 //Create new User 
 
-Route::post('/users', [UserController::class, 'store']);
+Route::post('/users', [UserController::class, 'store'])->middleware("guest");//only user store listing ;
+
+//log user out 
+
+Route::post('/logout', [UserController::class, 'logout'])->middleware("auth");//only user can logout ;
+
+//Show login form 
+Route::get('/login', [UserController::class, 'login'])->name("login")->middleware("guest");// add login page redirect route by name guest can use 
+
+// Log IN USer
+
+Route::post("/users/authenticate", [UserController::class, 'authenticate']);
